GroupTours/requirements.txt
asgiref==3.6.0
Brotli==1.1.0
click==8.2.1
colorama==0.4.6
dj-database-url==3.0.1
Django==4.2
django-cors-headers==4.7.0
django-filter==25.1
django-polymorphic==4.1.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
drf-polymorphic==2.1.1
gunicorn==23.0.0
h11==0.16.0
openpyxl==3.1.5
packaging==25.0
pi==0.1.2
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.0
pytz==2023.3
reportlab==4.4.4
qrcode==8.2
sqlparse==0.4.3
typing_extensions==4.14.1
tzdata==2023.3
uvicorn==0.35.0
whitenoise==6.9.0

# Testing y datos dummy
factory-boy==3.3.3
faker==38.2.0
